build: add 'examples' build step that supports multiple entries
Move out the main.zig file at the sdk root and into the example folder.
Move up all the entire psp/sdk folder.
Add a 'pspsdk' module that's easy to depend on for an external project.
Port clearScreen example to the updated sdk and make it build alongside
the ziggy cude example.