Remove safety hacks in sceGu and fix underlying bugs
Calls to `@setRuntimeSafety(false)` are harming correctness, and break in
all kinds of unexpected ways. The sane approach is to remove all those
calls and fix the underlying issues, which is why this commit exists.

This restores functionality of all examples that do graphics.
It also fixes Debug and ReleaseFast modes, which were broken before.

The API surface of `sceGu` is quite large, so there might still be issues
hiding in some corners, but they will be caught by @panic() and printed
on the debug screen. Sadly the debug screen is not a reliable way to
catch crashes anyway, but it's better than nothing.

As of right now the custom panic handler sleeps for 10 seconds before
exiting, so if you're running your homebrew with PPSSPP - you can look
out for sceKernelSomethingSomethingDelay() with a corresponding delay,
and you'll know that your program crashes this way.